port to 1.20
known issues:
alpha style logo doesn't work
many other things I didn't test
possibly forge since I wasn't able to test on forge

note, appleskin hasn't updated to forge 1.20 as of 23-06-08

not sure if it was necessary/or is appropriate to use CheckReturnValue annotation as I thought it could be a replacement for CheckForNull.

 I guess we could work from here, especially considering 9.0 so far was only new config ui.

logo outline tweak was removed due to it being unnecessary, however if there is demand it could be brought back. also the old logo rendering code shouldn't be brought back since the old logo can be brought back with a resource pack.